% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{combine_terms}
\alias{combine_terms}
\title{Create right-hand side of GAM formula by combining terms}
\usage{
combine_terms(vars, kval)
}
\description{
Create right-hand side of GAM formula by combining terms
}
\keyword{internal}
