% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{create_formulas}
\alias{create_formulas}
\title{Function to create GAM formulas
Gets all possible combinations of predictors, then combines them with the response variable
Next, these are converted into \code{formulas}.}
\usage{
create_formulas(response, vars, kval, max_n)
}
\description{
Function to create GAM formulas
Gets all possible combinations of predictors, then combines them with the response variable
Next, these are converted into \code{formulas}.
}
\keyword{internal}
