% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{examine_gam_res}
\alias{examine_gam_res}
\title{Plot diagnostics}
\usage{
examine_gam_res(g, data)
}
\arguments{
\item{g}{A \link[mgcv]{gam}}

\item{data}{Data frame used to produce \code{g}}
}
\description{
\code{examine_gam_res} produces diagnostic plots used to assess the validity of model assumptions.
Three figures are produced that show the following:
\enumerate{
  \item{Theoretical & observed quantiles}
  \item{Autocorrelation}
  \item{Residuals over time}
}
}
