% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{get_all_RHS}
\alias{get_all_RHS}
\title{Determine all possible combinations of terms for a GAM up to 3 terms}
\usage{
get_all_RHS(vars, max_n)
}
\description{
Determine all possible combinations of terms for a GAM up to 3 terms
}
\keyword{internal}
