% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_table_formula}
\alias{get_table_formula}
\title{Create table of GAM data from given list}
\usage{
get_table_formula(x)
}
\description{
Create table of GAM data from given list
}
\keyword{internal}
