% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_gam_effect}
\alias{plot_gam_effect}
\title{Plot effect of predictor for a GAM}
\usage{
plot_gam_effect(g, ylabel)
}
\arguments{
\item{g}{A \link[mgcv]{gam}}

\item{ylabel}{Label for the y axis (i.e., the name of the reponse variable)}
}
\description{
Produce a partial residual plot for each predictor of a \link[mgcv]{gam}. Plots are
on a 2\eqn{\times}{x}2 grid. Ribbons show the 95\% confidence intervals.
}
