% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_gam_ts}
\alias{plot_gam_ts}
\title{Plot time series of a GAM}
\usage{
plot_gam_ts(g, data, response)
}
\arguments{
\item{g}{A \link[mgcv]{gam}}

\item{data}{Data frame used to produce \code{g}}

\item{response}{Name of the response variable}
}
\description{
Plots figures to examine model results and observations over time, as well as the contribution of each predictor to the model.
Two figures are produced:
\enumerate{
  \item{Contributions of each predictor to the model over time}
  \item{Time series of observations & model fit with 95\% confidence interval}
}
}
