% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{rank_gams}
\alias{rank_gams}
\title{Extract top \code{n} GAMs from list according to AIC}
\usage{
rank_gams(gam_list, n)
}
\arguments{
\item{gam_list}{List of \link[mgcv]{gam}s}

\item{n}{Number of \link[mgcv]{gam}s to return}
}
\value{
List of \link[mgcv]{gam}s
}
\description{
A list of \link[mgcv]{gam}s are sorted according to \link[stats]{AIC} value.
\code{n} models with the lowest AIC are returned.
}
