% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{run_gams}
\alias{run_gams}
\title{Run GAM with given predictors & response variable on given data}
\usage{
run_gams(preds, response, kval, max_n, data)
}
\arguments{
\item{preds}{Names of the predictors (vector of strings)}

\item{response}{Name of response variable (string)}

\item{kval}{Number of knots in the smoothers (numeric)}

\item{max_n}{Maximum number of terms in each GAM (numeric)}

\item{data}{Data set that includes the predictors and response variable (data frame)}
}
\value{
List of \link[mgcv]{gam}s
}
\description{
Generalised additive models describing the response variable, \code{response}, are constructed using all
possible combinations of the given predictors, \code{preds}. These models are fit using the given data frame,
\code{data}, and contain up to and including \code{max_n} terms. These are smoothed terms with \code{k} knots.
}
