% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{tabulate_gams}
\alias{tabulate_gams}
\title{Create a data frame for a table}
\usage{
tabulate_gams(gam_list)
}
\arguments{
\item{gam_list}{List of \link[mgcv]{gam}s to be tabulated}
}
\value{
\link[knitr]{kable} table of \link[mgcv]{gam}s
}
\description{
Output a list of \link[mgcv]{gam}s in a \link[knitr]{kable} format showing model formula,
AIC, percent deviance explained, and GCV.
}
